config.py
"""
Configuration file for Stock Analysis Project
Contains all constants, database credentials, and the engineered prompt
"""

import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Stock symbols to analyze
STOCK_SYMBOLS = [
    "AAPL", "MSFT", "TSLA", "BABA", "SAP", 
    "NESN.SW", "AMZN", "TM", "SHEL", "NFLX", 
    "ASML", "SIE.DE", "NVO", "TCS.NS", "SHOP"
]

# Database configuration - Updated for local test database
DB_CONFIG = {
    'host': 'localhost',
    'port': 5432,
    'database': 'yfin_try',
    'user': 'postgres',
    'password': os.getenv('DB_PASSWORD', 'klaseba123')
}

# Hugging Face configuration (keeping for compatibility)
HUGGINGFACE_API_KEY = os.getenv('HUGGINGFACE_API_KEY')
HUGGINGFACE_MODEL = "microsoft/DialoGPT-medium"
HUGGINGFACE_API_URL = f"https://api-inference.huggingface.co/models/{HUGGINGFACE_MODEL}"

# Groq configuration (free LLM) - ACTIVELY USED
GROQ_API_KEY = os.getenv('GROQ_API_KEY')
USE_GROQ = True

# The engineered prompt
ANALYSIS_PROMPT = """
You are a financial analysis expert. I will provide you with a daily JSON data dump from the Yahoo Finance API for 10 predefined stocks. Your task is to analyze and summarize the data into concise, structured answers for each stock. These summaries will be displayed on a non-interactive financial dashboard, so they must be informative and digestible at a glance.

The raw JSON data will be passed to you as-is, without preprocessing. You should extract only the necessary values from this JSON to answer each question. If a required data field is missing, acknowledge it concisely and continue. When helpful, you may incorporate other fields to support or explain the analysis.

For each stock, provide the output in the following structure:

Stock Name: {Insert Company Name from the data}  
Stock Symbol: {Insert Symbol from the data defined in our data warehouse}  

Question 1: What is the current performance of this stock compared to its recent historical trend?  
Answer 1: {Your answer}  

Question 2: Is this stock considered overvalued or undervalued based on current analyst targets and earnings data?  
Answer 2: {Your answer}  

Question 3: What are the key financial strengths or weaknesses of this company based on its latest financial statements?  
Answer 3: {Your answer}  

Question 4: What do recent insider and institutional activities suggest about confidence in this stock?  
Answer 4: {Your answer}  

Question 5: How does the stock's volatility and risk profile compare to the broader market (e.g., VIX, Dow Futures)?  
Answer 5: {Your answer}

You may refer to the following JSON fields to answer the questions, but you are not limited to them:
• history, regularMarketPrice, regularMarketChangePercent, regularMarketPreviousClose, fast_info, quarterly_earnings
• analyst_price_targets, earnings, eps_trend, eps_revisions, recommendations_summary
• balance_sheet, income_stmt, cash_flow, ttm_financials
• insider_transactions, institutional_holders, major_holders
• info, beta, risk, and macro indicators like ^VIX, YM=F, RTY=F

Instructions:
– Each answer should be concise (preferably ≤ 100 words) and provide key takeaways rather than technical detail.
– Use a professional tone appropriate for a financial research dashboard. Analytical and neutral with mild narrative flow.
– Predictive statements are allowed if grounded in evidence, but avoid speculation.
– If relevant, explain confidence or uncertainty behind your conclusion (e.g., "Based on a limited sample of earnings…").
– Do not compare one stock to others unless explicitly asked. Each stock is to be evaluated independently.

This output will be presented to end users without interaction. It must deliver value at a glance, highlight relevant insights, and avoid raw data dumps or excessive detail.
"""