requirements.txt
yfinance==0.2.28
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
pandas>=2.0.0
numpy>=1.24.0